Make some more progress
The weird compiler that Android had for Java 8 is gone with the new
version, so all those issues are fixed. Now I just have to keep moving.
There's a comment in FloorDrawerHandler explaining the next steps.
Although it's wrong; I can send a broadcast directly from the drawer to
the fragments. Problem solved.