Add Firebase Messaging
It works. So that's pretty good. I think I should probably add the
trades page proper so that I can have something to actually show when
you click on the notification, then I have to figure out how to make the
notification less tiny, and make it so that clicking on it takes you
directly to the trade, and I think I'm done. How hard can that be?